Update secrets, use new renderer
