+++ pkg/koyeb/secrets.go
@@ -1,15 +1,7 @@
 package koyeb
 
 import (
-	"bufio"
-	"context"
-	"fmt"
-	"os"
-	"strings"
-
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/manifoldco/promptui"
-	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
@@ -39,6 +31,7 @@ func NewSecretCmd() *cobra.Command {
 	getSecretCmd := &cobra.Command{
 		Use:   "get NAME",
 		Short: "Get secret",
+		Args:  cobra.ExactArgs(1),
 		RunE:  h.Get,
 	}
 	secretCmd.AddCommand(getSecretCmd)
@@ -53,6 +46,7 @@ func NewSecretCmd() *cobra.Command {
 	describeSecretCmd := &cobra.Command{
 		Use:   "describe NAME",
 		Short: "Describe secrets",
+		Args:  cobra.ExactArgs(1),
 		RunE:  h.Describe,
 	}
 	secretCmd.AddCommand(describeSecretCmd)
@@ -87,212 +81,4 @@ func NewSecretHandler() *SecretHandler {
 }
 
 type SecretHandler struct {
-}
-
-func (h *SecretHandler) Create(cmd *cobra.Command, args []string, createSecret *koyeb.CreateSecret) error {
-	client := getApiClient()
-	ctx := getAuth(context.Background())
-
-	createSecret.SetName(args[0])
-	if !cmd.LocalFlags().Lookup("value-from-stdin").Changed && !cmd.LocalFlags().Lookup("value").Changed {
-		prompt := promptui.Prompt{
-			Label: "Enter your secret",
-			Mask:  '*',
-		}
-
-		result, err := prompt.Run()
-		if err != nil {
-			er(err)
-		}
-		createSecret.SetValue(result)
-	}
-	if cmd.LocalFlags().Lookup("value-from-stdin").Changed && cmd.LocalFlags().Lookup("value").Changed {
-		log.Fatalf("Cannot use value and value-from-stdin at the same time")
-	}
-	if cmd.LocalFlags().Lookup("value-from-stdin").Changed {
-		var input []string
-
-		scanner := bufio.NewScanner(os.Stdin)
-		for {
-			scanner.Scan()
-			text := scanner.Text()
-			if len(text) != 0 {
-				input = append(input, text)
-			} else {
-				break
-			}
-		}
-
-		createSecret.SetValue(strings.Join(input, "\n"))
-	}
-	_, _, err := client.SecretsApi.CreateSecret(ctx).Body(*createSecret).Execute()
-	if err != nil {
-		fatalApiError(err)
-	}
-	return nil
-}
-
-func (h *SecretHandler) Update(cmd *cobra.Command, args []string, updateSecret *koyeb.Secret) error {
-	client := getApiClient()
-	ctx := getAuth(context.Background())
-
-	if cmd.LocalFlags().Lookup("value-from-stdin").Changed && cmd.LocalFlags().Lookup("value").Changed {
-		log.Fatalf("Cannot use value and value-from-stdin at the same time")
-	}
-	if cmd.LocalFlags().Lookup("value-from-stdin").Changed {
-		var input []string
-
-		scanner := bufio.NewScanner(os.Stdin)
-		for {
-			scanner.Scan()
-			text := scanner.Text()
-			if len(text) != 0 {
-				input = append(input, text)
-			} else {
-				break
-			}
-		}
-
-		updateSecret.SetValue(strings.Join(input, "\n"))
-	}
-	_, _, err := client.SecretsApi.UpdateSecret2(ctx, args[0]).Body(*updateSecret).Execute()
-	if err != nil {
-		fatalApiError(err)
-	}
-	return nil
-}
-
-func (h *SecretHandler) Get(cmd *cobra.Command, args []string) error {
-	format := getFormat("table")
-	if len(args) == 0 {
-		return h.listFormat(cmd, args, format)
-	}
-	return h.getFormat(cmd, args, format)
-}
-
-func (h *SecretHandler) Describe(cmd *cobra.Command, args []string) error {
-	format := getFormat("detail")
-	if len(args) == 0 {
-		return h.listFormat(cmd, args, format)
-	}
-	return h.getFormat(cmd, args, format)
-}
-
-func (h *SecretHandler) Delete(cmd *cobra.Command, args []string) error {
-	client := getApiClient()
-	ctx := getAuth(context.Background())
-
-	for _, arg := range args {
-		_, _, err := client.SecretsApi.DeleteSecret(ctx, arg).Execute()
-		if err != nil {
-			fatalApiError(err)
-		}
-	}
-	return nil
-}
-
-func (h *SecretHandler) List(cmd *cobra.Command, args []string) error {
-	format := getFormat("table")
-	return h.listFormat(cmd, args, format)
-}
-
-func (h *SecretHandler) getFormat(cmd *cobra.Command, args []string, format string) error {
-	client := getApiClient()
-	ctx := getAuth(context.Background())
-
-	for _, arg := range args {
-		res, _, err := client.SecretsApi.GetSecret(ctx, arg).Execute()
-		if err != nil {
-			fatalApiError(err)
-		}
-		render(format, &GetSecretReply{res})
-	}
-
-	return nil
-}
-
-func (h *SecretHandler) listFormat(cmd *cobra.Command, args []string, format string) error {
-	client := getApiClient()
-	ctx := getAuth(context.Background())
-
-	page := 0
-	offset := 0
-	limit := 100
-	for {
-		res, _, err := client.SecretsApi.ListSecrets(ctx).Limit(fmt.Sprintf("%d", limit)).Offset(fmt.Sprintf("%d", offset)).Execute()
-		if err != nil {
-			fatalApiError(err)
-		}
-		render(format, &ListSecretsReply{res})
-		page += 1
-		offset = page * limit
-		if int64(offset) >= res.GetCount() {
-			break
-		}
-	}
-
-	return nil
-}
-
-type GetSecretReply struct {
-	koyeb.GetSecretReply
-}
-
-func (a *GetSecretReply) MarshalBinary() ([]byte, error) {
-	return a.GetSecretReply.GetSecret().MarshalJSON()
-}
-
-func (a *GetSecretReply) Title() string {
-	return "Secret"
-}
-
-func (a *GetSecretReply) Headers() []string {
-	return []string{"id", "name", "type", "value", "updated_at"}
-}
-
-func (a *GetSecretReply) Fields() []map[string]string {
-	res := []map[string]string{}
-	item := a.GetSecret()
-	fields := map[string]string{}
-	for _, field := range a.Headers() {
-		if field == "value" {
-			fields[field] = "*****"
-		} else {
-			fields[field] = GetField(item, field)
-		}
-	}
-	res = append(res, fields)
-	return res
-}
-
-type ListSecretsReply struct {
-	koyeb.ListSecretsReply
-}
-
-func (a *ListSecretsReply) Title() string {
-	return "Secrets"
-}
-
-func (a *ListSecretsReply) MarshalBinary() ([]byte, error) {
-	return a.ListSecretsReply.MarshalJSON()
-}
-
-func (a *ListSecretsReply) Headers() []string {
-	return []string{"id", "name", "type", "value", "updated_at"}
-}
-
-func (a *ListSecretsReply) Fields() []map[string]string {
-	res := []map[string]string{}
-	for _, item := range a.GetSecrets() {
-		fields := map[string]string{}
-		for _, field := range a.Headers() {
-			if field == "value" {
-				fields[field] = "*****"
-			} else {
-				fields[field] = GetField(item, field)
-			}
-		}
-		res = append(res, fields)
-	}
-	return res
 }
